service-1/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
databases
psycopg2-binary
asyncpg
pika
opencv-python
boto3
requests